Cleaning up the `dbtables`, there were some issues in the move to the `helpers/dbtables` folder.
Removed the use of the `tableID` variable in the `dbtables` endpoint.
